Revert "Refactor Endpoints (#922)"
This reverts commit 46abf49699f26f4437ddc3e80de0531dc6842b65.
Resolve revert conflict
