+++ pkg/importer/tests-openapi/multiple-return-values.sysl
@@ -11,11 +11,7 @@ testapp "Simple" [package="package_foo"]:
     /test:
         GET:
             | No description.
-            return 400
-            return 401
-            return 500
-            return 503
-            return 504
+            return error
             return ok <: _test_ok
 
     #---------------------------------------------------------------------------
