+++ pkg/importer/tests-swagger/default-response.sysl
@@ -14,7 +14,7 @@ testapp "Goat CRUD API" [package="package_foo"]:
     /goat/status:
         GET:
             | Check goat status
-            return default
+            return error
 
     #---------------------------------------------------------------------------
     # definitions
