+++ pkg/importer/tests-swagger/error-response.sysl
@@ -14,8 +14,7 @@ testapp "Goat CRUD API" [package="package_foo"]:
     /goat/status:
         GET:
             | Check goat status
-            return 400
-            return 500
+            return error
             return ok
 
     #---------------------------------------------------------------------------
