+++ pkg/importer/tests-swagger/param-array.sysl
@@ -11,8 +11,7 @@ testapp "Simple" [package="package_foo"]:
     /simple:
         PUT (DocumentRequest <: Document [mediatype="application/json", ~body]):
             | No description.
-            return 400
-            return 401
+            return error
             return ok
 
     #---------------------------------------------------------------------------
