+++ pkg/importer/tests-swagger/x-dash-response.sysl
@@ -14,7 +14,7 @@ testapp "Goat CRUD API" [package="package_foo"]:
     /goat/status:
         GET:
             | Check goat status
-            return x-banana
+            return error
 
     #---------------------------------------------------------------------------
     # definitions
